Refactoring business logic to services folder
